Restructured the file structure a lot to make it easier to find specific parts and also add new ones
