+++ src/database/mod.rs
@@ -1,0 +1,3 @@
+mod get_credentials;
+
+pub use get_credentials::*;
