+++ src/kubernetes/mod.rs
@@ -1,0 +1,3 @@
+mod auth;
+
+pub use auth::*;
