+++ src/kv2/mod.rs
@@ -1,0 +1,19 @@
+mod configure;
+mod delete;
+mod delete_metadata_all_versions;
+mod delete_versions;
+mod destroy_versions;
+mod get;
+mod undelete_versions;
+mod update_set;
+
+pub use configure::*;
+pub use delete::*;
+pub use delete_metadata_all_versions::*;
+pub use delete_versions::*;
+pub use destroy_versions::*;
+pub use get::*;
+pub use undelete_versions::*;
+pub use update_set::*;
+
+// TODO: Add List, Read Metadata, Update Metadata
